+++ upload/admin/controller/extension/payment/globalpayments_affirm.php
@@ -22,12 +22,17 @@ class ControllerExtensionPaymentGlobalPaymentsAffirm extends Controller {
 		} else {
 			$data['payment_globalpayments_affirm_title'] = $this->config->get('payment_globalpayments_affirm_title');
 		}
-
 		if (isset($this->request->post['payment_globalpayments_affirm_payment_action'])) {
 			$data['payment_globalpayments_affirm_payment_action'] = $this->request->post['payment_globalpayments_affirm_payment_action'];
 		} else {
 			$data['payment_globalpayments_affirm_payment_action'] = $this->config->get('payment_globalpayments_affirm_payment_action');
 		}
+		if (isset($this->request->post['payment_globalpayments_affirm_sort_order'])) {
+			$data['payment_globalpayments_affirm_sort_order'] = $this->request->post['payment_globalpayments_affirm_sort_order'];
+		} else {
+			$data['payment_globalpayments_affirm_sort_order'] = $this->config->get('payment_globalpayments_affirm_sort_order') ?? 0;
+		}
+
 		$data['active_tab'] = str_replace('extension/payment/globalpayments_', '', $this->request->get['route']);
 
 		return $this->load->view('extension/payment/globalpayments_affirm', $data);