+++ upload/admin/language/en-gb/extension/payment/globalpayments_affirm.php
@@ -3,12 +3,13 @@
 $_['heading_title'] = 'GlobalPayments - Affirm';
 
 // Text
-$_['text_globalpayments_affirm'] = '<a href="https://developer.globalpay.com" target="_blank"><img src="https://developer.globalpay.com/static/media/logo.dab7811d.svg" alt="Global Payments" title="Global Payments" style="border: 1px solid #EEEEEE;"></a>';
+$_['text_globalpayments_affirm'] = '<a href="https://developer.globalpay.com" target="_blank"><img src="https://avatars.githubusercontent.com/u/25797248?s=200&v=4" width="40px" height="40px" alt="Global Payments" title="Global Payments" style="border: 1px solid #EEEEEE;"></a>';
 
 // Label
 $_['label_enabled']                = 'Enable/Disable';
 $_['label_title']                  = 'Title';
 $_['label_payment_action']         = 'Payment Action';
+$_['label_sort_order']             = 'Sort Order';
 
 // Help
 $_['help_title']          = 'This controls the title which the user sees during checkout.';