+++ upload/admin/language/en-gb/extension/payment/globalpayments_ucp.php
@@ -11,12 +11,10 @@ $_['tab_clicktopay']     = 'Click To Pay';
 $_['tab_affirm']         = 'Affirm';
 $_['tab_klarna']         = 'Klarna';
 $_['tab_clearpay']       = 'Clearpay';
-$_['tab_sepa']           = 'Sepa';
-$_['tab_fasterpayments'] = 'Faster Payments';
-
+$_['tab_openbanking']    = 'Bank Payment';
 
 // Text
-$_['text_globalpayments_ucp'] = '<a href="https://developer.globalpay.com" target="_blank"><img src="https://developer.globalpay.com/static/media/logo.dab7811d.svg" alt="Global Payments" title="Global Payments" style="border: 1px solid #EEEEEE;"></a>';
+$_['text_globalpayments_ucp'] = '<a href="https://developer.globalpay.com" target="_blank"><img src="https://avatars.githubusercontent.com/u/25797248?s=200&v=4" width="40px" height="40px" alt="Global Payments" title="Global Payments" style="border: 1px solid #EEEEEE;"></a>';
 $_['text_extension']          = 'Extensions';
 $_['text_edit']               = 'Edit GlobalPayments - Unified Payments';
 $_['text_success']            = 'Success: You have modified GlobalPayments - Unified Payments account details!';
@@ -38,6 +36,7 @@ $_['label_payment_action']        = 'Payment Action';
 $_['label_allow_card_saving']     = 'Allow Card Saving';
 $_['label_txn_descriptor']        = 'Order Transaction Descriptor';
 $_['label_enable_three_d_secure'] = 'Enable 3DSecure';
+$_['label_sort_order']            = 'Sort Order';
 
 // Help
 $_['help_title']                 = 'This controls the title which the user sees during checkout.';