+++ upload/catalog/language/en-gb/extension/payment/globalpayments_fasterpayments.php
@@ -1,9 +1,0 @@
-<?php
-// Placeholder
-$_['placeholder_title'] = 'FasterPayments';
-
-// Text
-$_['text_paymentMethod_not_supported'] = 'FasterPayments is not supported on this device/browser for the specified country - currency. Please choose another payment method.';
-$_['text_invalidCustomerDetails'] = 'FasterPayments requires adress postcode and phone number to be filled in your account data.';
-$_['text_paymentMethod_processing_default_error'] = 'An error occurred while processing the payment. Please try again or use a different payment method.';
-$_['text_paymentMethod_processing_failed_error'] = 'Your payment was unsuccessful. Please try again or use a different payment method.';