+++ upload/system/library/globalpayments.php
@@ -20,14 +20,13 @@ use GlobalPayments\PaymentGatewayProvider\PaymentMethods\DigitalWallets\GooglePa
 use GlobalPayments\PaymentGatewayProvider\PaymentMethods\BuyNowPayLater\Affirm;
 use GlobalPayments\PaymentGatewayProvider\PaymentMethods\BuyNowPayLater\Klarna;
 use GlobalPayments\PaymentGatewayProvider\PaymentMethods\BuyNowPayLater\Clearpay;
-use GlobalPayments\PaymentGatewayProvider\PaymentMethods\OpenBanking\FasterPayments;
-use GlobalPayments\PaymentGatewayProvider\PaymentMethods\OpenBanking\Sepa;
+use GlobalPayments\PaymentGatewayProvider\PaymentMethods\OpenBanking\OpenBanking;
 
 class GlobalPayments {
 	/**
 	 * Extension version.
 	 */
-	const VERSION = '1.6.0';
+	const VERSION = '1.6.1';
 
 	/**
 	 * @var GlobalPayments\PaymentGatewayProvider\Gateways\GatewayInterface
@@ -78,11 +77,8 @@ class GlobalPayments {
 			case Clearpay::PAYMENT_METHOD_ID:
 				$this->setClearpayPaymentMethod();
 				break;
-			case Sepa::PAYMENT_METHOD_ID:
-				$this->setSepaPaymentMethod();
-				break;
-			case FasterPayments::PAYMENT_METHOD_ID:
-				$this->setFasterPaymentsPaymentMethod();
+			case OpenBanking::PAYMENT_METHOD_ID:
+				$this->setOpenBankingPaymentMethod();
 				break;
 		}
 	}
@@ -231,25 +227,17 @@ class GlobalPayments {
 		$this->gateway->logDirectory             = DIR_LOGS;
 	}
 
-	public function setSepaPaymentMethod() {
-		$this->paymentMethod                = new Sepa($this->gateway);
-		$this->paymentMethod->enabled       = $this->config->get('payment_globalpayments_sepa_enabled');
-		$this->paymentMethod->title         = $this->config->get('payment_globalpayments_sepa_title');
-		$this->paymentMethod->paymentAction = $this->config->get('payment_globalpayments_sepa_payment_action');
-		$this->paymentMethod->countries     = $this->config->get('payment_globalpayments_sepa_countries');
-		$this->paymentMethod->iban          = $this->config->get('payment_globalpayments_sepa_iban');
-		$this->paymentMethod->accountName   = $this->config->get('payment_globalpayments_sepa_account_name');
-	}
-
-	public function setFasterPaymentsPaymentMethod() {
-		$this->paymentMethod                = new FasterPayments($this->gateway);
-		$this->paymentMethod->enabled       = $this->config->get('payment_globalpayments_fasterpayments_enabled');
-		$this->paymentMethod->title         = $this->config->get('payment_globalpayments_fasterpayments_title');
-		$this->paymentMethod->paymentAction = $this->config->get('payment_globalpayments_fasterpayments_payment_action');
-		$this->paymentMethod->countries     = $this->config->get('payment_globalpayments_fasterpayments_countries');
-		$this->paymentMethod->sortCode      = $this->config->get('payment_globalpayments_fasterpayments_sort_code');
-		$this->paymentMethod->accountName   = $this->config->get('payment_globalpayments_fasterpayments_account_name');
-		$this->paymentMethod->accountNumber = $this->config->get('payment_globalpayments_fasterpayments_account_number');
+	public function setOpenBankingPaymentMethod() {
+		$this->paymentMethod                = new OpenBanking($this->gateway);
+		$this->paymentMethod->enabled       = $this->config->get('payment_globalpayments_openbanking_enabled');
+		$this->paymentMethod->title         = $this->config->get('payment_globalpayments_openbanking_title');
+		$this->paymentMethod->paymentAction = $this->config->get('payment_globalpayments_openbanking_payment_action');
+		$this->paymentMethod->countries     = $this->config->get('payment_globalpayments_openbanking_countries');
+		$this->paymentMethod->sortCode      = $this->config->get('payment_globalpayments_openbanking_sort_code');
+		$this->paymentMethod->accountName   = $this->config->get('payment_globalpayments_openbanking_account_name');
+		$this->paymentMethod->accountNumber = $this->config->get('payment_globalpayments_openbanking_account_number');
+		$this->paymentMethod->iban          = $this->config->get('payment_globalpayments_openbanking_iban');
+		$this->paymentMethod->currencies    = $this->config->get('payment_globalpayments_openbanking_currencies');
 	}
 
 	public function setSecurePaymentFieldsTranslations() {